move ut and add ut for schema less
